LPS-114698 Move
